Add table to projects page
